refactor : 구조 리팩 토링
