+++ DuDoong-Infrastructure/src/main/java/band/gosrock/infrastructure/config/ses/AwsSesUtils.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Component;
 import org.thymeleaf.context.Context;
 import org.thymeleaf.spring5.SpringTemplateEngine;
 import software.amazon.awssdk.services.ses.SesAsyncClient;
+import software.amazon.awssdk.services.ses.SesClient;
 import software.amazon.awssdk.services.ses.model.Body;
 import software.amazon.awssdk.services.ses.model.Content;
 import software.amazon.awssdk.services.ses.model.Destination;
@@ -19,20 +20,18 @@ import software.amazon.awssdk.services.ses.model.SendEmailResponse;
 @AllArgsConstructor
 public class AwsSesUtils {
 
-    private final SesAsyncClient sesAsyncClient;
+    private final SesClient sesClient;
     private final SpringTemplateEngine templateEngine;
 
-    public void singleEmailRequest(String to, String subject, String template, Context context)
-        throws ExecutionException, InterruptedException {
+    public void singleEmailRequest(String to, String subject, String template, Context context){
         String html = templateEngine.process(template, context);
 
         Builder sendEmailRequestBuilder = SendEmailRequest.builder();
         sendEmailRequestBuilder.destination(Destination.builder().toAddresses(to).build());
         sendEmailRequestBuilder.message(newMessage(subject, html)).source("life@dudoong.com").build();
 
-        CompletableFuture<SendEmailResponse> sendEmailResponseCompletableFuture = sesAsyncClient.sendEmail(
+        SendEmailResponse sendEmailResponse1 = sesClient.sendEmail(
             sendEmailRequestBuilder.build());
-        SendEmailResponse sendEmailResponse = sendEmailResponseCompletableFuture.get();
     }
 
     private Message newMessage(String subject, String html) {
